+++ src/main/java/ru/vyarus/dropwizard/guice/test/jupiter/ext/TestGuiceyAppExtension.java
@@ -18,10 +18,13 @@ import ru.vyarus.dropwizard.guice.test.jupiter.ext.conf.TestExtensionsTracker;
 import ru.vyarus.dropwizard.guice.test.util.ConfigOverrideUtils;
 import ru.vyarus.dropwizard.guice.test.util.ConfigurablePrefix;
 import ru.vyarus.dropwizard.guice.test.util.HooksUtil;
+import ru.vyarus.dropwizard.guice.test.util.ReusableAppUtils;
 import ru.vyarus.dropwizard.guice.test.util.TestSetupUtils;
 
+import java.lang.reflect.AnnotatedElement;
 import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 
 /**
  * {@link TestGuiceyApp} junit 5 extension implementation. Normally, extension should be activated with annotation,
@@ -94,17 +97,16 @@ public class TestGuiceyAppExtension extends GuiceyExtensionsSupport {
     }
 
     @Override
-    protected DropwizardTestSupport<?> prepareTestSupport(final String configPrefix,
-                                                          final ExtensionContext context,
-                                                          final List<TestEnvironmentSetup> setups) {
+    protected ExtensionConfig getConfig(final ExtensionContext context) {
         if (config == null) {
             // Configure from annotation
             // Note that it is impossible to have both manually build config and annotation because annotation
             // will be processed first and manual registration will be simply ignored
 
+            final Optional<AnnotatedElement> element = context.getElement();
             final TestGuiceyApp ann = AnnotationSupport
                     // also search annotation inside other annotations (meta)
-                    .findAnnotation(context.getElement(), TestGuiceyApp.class).orElse(null);
+                    .findAnnotation(element, TestGuiceyApp.class).orElse(null);
 
             // catch incorrect usage by direct @ExtendWith(...)
             Preconditions.checkNotNull(ann, "%s annotation not declared: can't work without configuration, "
@@ -112,8 +114,22 @@ public class TestGuiceyAppExtension extends GuiceyExtensionsSupport {
                     TestGuiceyApp.class.getSimpleName(),
                     RegisterExtension.class.getSimpleName());
             config = Config.parse(ann, tracker);
+            if (config.reuseApp) {
+                // identify annotation extension source (class) and validate declaration correctness
+                ReusableAppUtils.registerAnnotation(context.getRequiredTestClass(), ann, config);
+            }
         }
+        if (config.reuseApp && config.reuseSource == null) {
+            // identify manual extension source (field) and validate declaration correctness
+            ReusableAppUtils.registerField(context.getRequiredTestClass(), this, config);
+        }
+        return config;
+    }
 
+    @Override
+    protected DropwizardTestSupport<?> prepareTestSupport(final String configPrefix,
+                                                          final ExtensionContext context,
+                                                          final List<TestEnvironmentSetup> setups) {
         // setups from @EnableSetup fields go last
         config.extensions.addAll(setups);
         TestSetupUtils.executeSetup(config, context);
@@ -268,6 +284,7 @@ public class TestGuiceyAppExtension extends GuiceyExtensionsSupport {
             res.hooksFromAnnotation(ann.annotationType(), ann.hooks());
             res.extensionsFromAnnotation(ann.annotationType(), ann.setup());
             res.tracker.debug = ann.debug();
+            res.reuseApp = ann.reuseApplication();
             return res;
         }
     }